+++ test/integration/elastic_db_repo/run_program.py
@@ -17,13 +17,13 @@
 import sys
 import os
 import shutil
-import time
 import unittest
 
 # Local
 sys.path.append(os.getcwd())
 import elastic_db_repo                          # pylint:disable=E0401,C0413
 import lib.gen_libs as gen_libs             # pylint:disable=E0401,C0413,R0402
+import lib.gen_class as gen_class           # pylint:disable=E0401,C0413,R0402
 import elastic_lib.elastic_class as elcs    # pylint:disable=E0401,C0413,R0402
 import version                                  # pylint:disable=E0401,C0413
 
@@ -77,16 +77,19 @@ class UnitTest(unittest.TestCase):
             "-S": elastic_db_repo.delete_dump,
             "-M": elastic_db_repo.rename_repo,
             "-U": elastic_db_repo.disk_usage}
-        self.args = {"-c": "elastic", "-d": self.config_path}
+        self.opt_val = ["-c", "-d"]
+        self.argv = [
+            "elastic_db_repo.py", "-c", "elastic", "-d", self.config_path,
+            "-z"]
+        self.args = gen_class.ArgParser(self.argv, opt_val=self.opt_val)
+        self.args.arg_parse2()
         self.user = self.cfg.user if hasattr(self.cfg, "user") else None
         self.japd = self.cfg.japd if hasattr(self.cfg, "japd") else None
         self.ca_cert = self.cfg.ssl_client_ca if hasattr(
             self.cfg, "ssl_client_ca") else None
-        self.scheme = self.cfg.scheme if hasattr(
-            self.cfg, "scheme") else "https"
         self.els = elcs.ElasticSearchRepo(
-            self.cfg.host, port=self.cfg.port, user=self.user, japd=self.japd,
-            ca_cert=self.ca_cert, scheme=self.scheme)
+            self.cfg.host, user=self.user, japd=self.japd,
+            ca_cert=self.ca_cert)
         self.els.connect()
         self.els2 = None
 
@@ -113,9 +116,8 @@ class UnitTest(unittest.TestCase):
             self.skipTest("test_delete_dump: Pre-conditions not met.")
 
         els2 = elcs.ElasticSearchDump(
-            self.cfg.host, port=self.cfg.port, repo=self.repo_name,
-            user=self.user, japd=self.japd, ca_cert=self.ca_cert,
-            scheme=self.scheme)
+            self.cfg.host, repo=self.repo_name, user=self.user, japd=self.japd,
+            ca_cert=self.ca_cert)
         els2.connect()
         els2.dump_name = self.dump_name
         err_flag, msg = els2.dump_db()
@@ -125,11 +127,11 @@ class UnitTest(unittest.TestCase):
             print(f"Reason: {msg}")
             self.skipTest("Dump failed")
 
-        self.args["-S"] = self.dump_name
-        self.args["-r"] = self.repo_name
+        self.args.args_array["-S"] = self.dump_name
+        self.args.args_array["-r"] = self.repo_name
 
-        self.assertFalse(elastic_db_repo.run_program(
-            self.args, self.func_names))
+        self.assertFalse(
+            elastic_db_repo.run_program(self.args, self.func_names))
 
     def test_list_dumps(self):
 
@@ -150,9 +152,8 @@ class UnitTest(unittest.TestCase):
             self.skipTest("test_list_dumps: Pre-conditions not met.")
 
         els2 = elcs.ElasticSearchDump(
-            self.cfg.host, port=self.cfg.port, repo=self.repo_name,
-            user=self.user, japd=self.japd, ca_cert=self.ca_cert,
-            scheme=self.scheme)
+            self.cfg.host, repo=self.repo_name, user=self.user, japd=self.japd,
+            ca_cert=self.ca_cert)
         els2.connect()
         err_flag, msg = els2.dump_db()
 
@@ -161,13 +162,11 @@ class UnitTest(unittest.TestCase):
             print(f"Reason: {msg}")
             self.skipTest("Dump failed")
 
-        self.args["-L"] = self.repo_name
+        self.args.args_array["-L"] = self.repo_name
 
-        with gen_libs.no_std_out():
-            self.assertFalse(
-                elastic_db_repo.run_program(self.args, self.func_names))
+        self.assertFalse(
+            elastic_db_repo.run_program(self.args, self.func_names))
 
-    @unittest.skip("Error:  Fails in a docker setup environment.")
     def test_disk_usage(self):
 
         """Function:  test_disk_usage
@@ -186,19 +185,10 @@ class UnitTest(unittest.TestCase):
             print(f"Reason:  {status_msg}")
             self.skipTest("test_disk_usage: Pre-conditions not met.")
 
-        # Wait until the repo dir has been created.
-        while True:
-            if not os.path.isdir(self.phy_repo_dir):
-                time.sleep(1)
+        self.args.args_array["-U"] = True
 
-            else:
-                break
-
-        self.args["-U"] = True
-
-        with gen_libs.no_std_out():
-            self.assertFalse(
-                elastic_db_repo.run_program(self.args, self.func_names))
+        self.assertFalse(
+            elastic_db_repo.run_program(self.args, self.func_names))
 
     def test_rename_repo(self):
 
@@ -218,13 +208,13 @@ class UnitTest(unittest.TestCase):
             print(f"Reason:  {status_msg}")
             self.skipTest("test_rename_repo: Pre-conditions not met.")
 
-        self.args["-M"] = [self.repo_name, self.repo_name2]
+        self.args.args_array["-M"] = [self.repo_name, self.repo_name2]
 
         elastic_db_repo.run_program(self.args, self.func_names)
 
         self.els2 = elcs.ElasticSearchRepo(
-            self.cfg.host, self.cfg.port, repo=self.repo_name2, user=self.user,
-            japd=self.japd, ca_cert=self.ca_cert, scheme=self.scheme)
+            self.cfg.host, repo=self.repo_name2, user=self.user,
+            japd=self.japd, ca_cert=self.ca_cert)
         self.els2.connect()
 
         self.assertIn(self.repo_name2, self.els2.repo_dict)
@@ -247,11 +237,10 @@ class UnitTest(unittest.TestCase):
             print(f"Reason:  {status_msg}")
             self.skipTest("test_list_repos: Pre-conditions not met.")
 
-        self.args["-R"] = True
+        self.args.args_array["-R"] = True
 
-        with gen_libs.no_std_out():
-            self.assertFalse(
-                elastic_db_repo.run_program(self.args, self.func_names))
+        self.assertFalse(
+            elastic_db_repo.run_program(self.args, self.func_names))
 
     def test_delete_repo(self):
 
@@ -271,13 +260,13 @@ class UnitTest(unittest.TestCase):
             print(f"Reason:  {status_msg}")
             self.skipTest("test_list_repos: Pre-conditions not met.")
 
-        self.args["-D"] = self.repo_name
+        self.args.args_array["-D"] = self.repo_name
 
         elastic_db_repo.run_program(self.args, self.func_names)
 
         self.els2 = elcs.ElasticSearchRepo(
-            self.cfg.host, self.cfg.port, repo=self.repo_name, user=self.user,
-            japd=self.japd, ca_cert=self.ca_cert, scheme=self.scheme)
+            self.cfg.host, repo=self.repo_name, user=self.user,
+            japd=self.japd, ca_cert=self.ca_cert)
         self.els2.connect()
 
         self.assertNotIn(self.repo_name, self.els2.repo_dict)
@@ -292,14 +281,14 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        self.args["-C"] = self.repo_name
-        self.args["-l"] = self.cfg.log_repo_dir
+        self.args.args_array["-C"] = self.repo_name
+        self.args.args_array["-l"] = self.cfg.log_repo_dir
 
         elastic_db_repo.run_program(self.args, self.func_names)
 
         self.els2 = elcs.ElasticSearchRepo(
-            self.cfg.host, self.cfg.port, repo=self.repo_name, user=self.user,
-            japd=self.japd, ca_cert=self.ca_cert, scheme=self.scheme)
+            self.cfg.host, repo=self.repo_name, user=self.user,
+            japd=self.japd, ca_cert=self.ca_cert)
         self.els2.connect()
 
         self.assertIn(self.repo_name, self.els2.repo_dict)
@@ -314,11 +303,12 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        if "-C" in self.args or "-R" in self.args or "-U" in self.args \
-           or "-L" in self.args or "-S" in self.args:
+        if "-C" in self.args.args_array or "-R" in self.args.args_array \
+           or "-U" in self.args.args_array or "-L" in self.args.args_array \
+           or "-S" in self.args.args_array:
             els = elcs.ElasticSearchRepo(
-                self.cfg.host, port=self.cfg.port, user=self.user,
-                japd=self.japd, ca_cert=self.ca_cert, scheme=self.scheme)
+                self.cfg.host, user=self.user, japd=self.japd,
+                ca_cert=self.ca_cert)
             els.connect()
 
             err_flag, status_msg = els.delete_repo(self.repo_name)
@@ -327,10 +317,10 @@ class UnitTest(unittest.TestCase):
                 print(f"Error: Failed to remove repository {self.repo_name}")
                 print(f"Reason:  {status_msg}")
 
-        elif "-M" in self.args:
+        elif "-M" in self.args.args_array:
             els = elcs.ElasticSearchRepo(
-                self.cfg.host, port=self.cfg.port, user=self.user,
-                japd=self.japd, ca_cert=self.ca_cert, scheme=self.scheme)
+                self.cfg.host, user=self.user, japd=self.japd,
+                ca_cert=self.ca_cert)
             els.connect()
 
             err_flag, status_msg = els.delete_repo(self.repo_name2)