+++ CHANGELOG.md
@@ -3,6 +3,39 @@ All notable changes to this project will be documented in this file.
 
 The format is based on "Keep a Changelog".  This project adheres to Semantic Versioning.
 
+
+## [2.0.0] - 2025-09-12
+Breaking Change
+
+- Converted output to JSON format for the dump list, repo list and disk partition usage options.
+- Added data out options (i.e. email, file, standard out, suppression) to the dump list, repo list and disk partition usage options.
+- Updated python-lib v4.0.3
+- Updated elastic-lib v4.2.0
+
+### Fixed:
+- delete_dump: Added ignore=False to get_dump_list call.
+
+### Added
+- get_dumps: Retrieve dumps from the Elasticsearch cluster and return the dumps in a dictionary format.
+- create_header: Create standard dictionary header and add Check entry to header if needed.
+- data_out: Determine where the data will be sent to such as email, file, standard out.
+
+### Changed
+- delete_dump: Converted from a single dump to a list and changed get_dump_list from a function call to a method call.
+- main: Added options for the data_out function to the checks, added -S to multiple values list.
+- disk_usage, list_dumps, list_repos: Refactored function and converted output to JSON format and added call to create_header and data_out.
+- run_program: Initialized an instance of the gen_class.TimeFormat class and passed to all functions called.
+- run_program: Removed port and scheme arguments to the elastic_class calls and removed the repo argument from elastic_class.ElasticSearchRepo instance call.
+- config/elastic.py.TEMPLATE: Removed port and scheme entries.
+- Documentation changes.
+
+### Deprecated
+- Support for Elasticsearch v7.4, v7.12 and v7.17
+
+### Removed
+- Support for Elasticsearch v6.4
+
+
 ## [1.0.0] - 2025-02-25
 - General Release.
 - Breaking Changes